pyproject.toml
[project]
name = "ustudy-unicraft"
version = "0.1.0"
description = ""
authors = [
    {name = "bahodir",email = "vbahodir00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "amqp (==5.3.1)",
    "argon2-cffi (==25.1.0)",
    "argon2-cffi-bindings (==25.1.0)",
    "asgiref (==3.9.1)",
    "attrs (==25.3.0)",
    "autobahn (==24.4.2)",
    "automat (==25.4.16)",
    "billiard (==4.2.1)",
    "boto3 (==1.40.9)",
    "botocore (==1.40.9)",
    "celery (==5.5.3)",
    "celery-types (==0.23.0)",
    "certifi (==2025.7.14)",
    "cfgv (==3.4.0)",
    "channels (==4.3.1)",
    "channels-redis (==4.3.0)",
    "charset-normalizer (==3.4.2)",
    "click (==8.2.1)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1.2)",
    "click-repl (==0.3.0)",
    "colorama (==0.4.6)",
    "constantly (==23.10.4)",
    "cron-descriptor (==1.4.5)",
    "cryptography (==45.0.5)",
    "daphne (==4.2.1)",
    "distlib (==0.4.0)",
    "django (==5.2.4)",
    "django-celery-beat (==2.8.1)",
    "django-cors-headers (==4.7.0)",
    "django-extensions (==4.1)",
    "django-unfold (==0.8.0)",
    "django-query-counter (==0.4.1)",
    "django-redis (==6.0.0)",
    "django-storages (==1.14.6)",
    "django-stubs (==5.2.2)",
    "django-stubs-ext (==5.2.2)",
    "django-timezone-field (==7.1)",
    "djangorestframework (==3.16.0)",
    "djangorestframework-simplejwt (==5.5.1)",
    "drf-spectacular (==0.28.0)",
    "drf-yasg (==1.21.10)",
    "filelock (==3.18.0)",
    "gevent (==25.5.1)",
    "greenlet (==3.2.3)",
    "hyperlink (==21.0.0)",
    "identify (==2.6.12)",
    "idna (==3.10)",
    "incremental (==24.7.2)",
    "inflection (==0.5.1)",
    "jmespath (==1.0.1)",
    "jsonschema (==4.25.0)",
    "jsonschema-specifications (==2025.4.1)",
    "kombu (==5.5.4)",
    "minio (==7.2.16)",
    "msgpack (==1.1.1)",
    "nodeenv (==1.9.1)",
    "packaging (==25.0)",
    "pillow (==11.3.0)",
    "platformdirs (==4.3.8)",
    "pre-commit (==4.2.0)",
    "prompt-toolkit (==3.0.51)",
    "psycopg2-binary (==2.9.10)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pycparser (==2.22)",
    "pycryptodome (==3.23.0)",
    "pyjwt (==2.10.1)",
    "pyopenssl (==25.1.0)",
    "python-crontab (==3.3.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-decouple (==3.8)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "redis (==6.2.0)",
    "referencing (==0.36.2)",
    "requests (==2.32.4)",
    "rpds-py (==0.26.0)",
    "s3transfer (==0.13.1)",
    "service-identity (==24.2.0)",
    "setuptools (==80.9.0)",
    "six (==1.17.0)",
    "sqlparse (==0.5.3)",
    "tabulate (==0.9.0)",
    "twisted (==25.5.0)",
    "txaio (==25.6.1)",
    "types-pyyaml (==6.0.12.20250516)",
    "typing-extensions (==4.14.1)",
    "tzdata (==2025.2)",
    "uritemplate (==4.2.0)",
    "urllib3 (==2.5.0)",
    "vine (==5.1.0)",
    "virtualenv (==20.32.0)",
    "wcwidth (==0.2.13)",
    "whitenoise (==6.9.0)",
    "zope-event (==5.1.1)",
    "zope-interface (==7.2)",
    "pandas (==2.3.2)",
    "numpy (==2.3.2)",
    "openpyxl (==3.1.5)",
    "et-xmlfile (==2.0.0)",
    "faker (==37.6.0)",
    "django-filter (==25.1)",
    "django-ckeditor (==6.7.3)",
    "django-js-asset (==3.1.2)",
    "django-cleanup (==9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
